[Move] DependencyValues 관련 코드 이동
extension 모아서 파일 구성 X
각각의 파일에서 볼 수 있도록 함.
( 코드 가독성 )
