+++ Moment_TCA_1.9.2/Moment/Moment/Service/SwiftDataService.swift
@@ -95,3 +95,12 @@ extension SwiftDataService: DependencyKey {
         }
     )
 }
+
+// MARK: - TCA : DependencyValues +
+extension DependencyValues {
+    // SwiftData Service
+    var swiftDataService: SwiftDataService {
+        get { self[SwiftDataService.self] }
+        set { self[SwiftDataService.self] = newValue }
+    }
+}
